make main entire screen. removed flexbox.
my thinking at this stage is make the entire image the 'background', then use z-index to bring .right up so it's on top of the background image in the foreground. not sure if this will work but worth a try. i also used dvh instead of vh because of stackoverflow answer here: https://stackoverflow.com/questions/74144034/why-is-the-css-height100vh-rule-exceeding-the-viewport-height-on-mobile-device so that the height auto-adjusts for mobile state