build left-side of design (#1)
* add main wrapper for semantic html

* add heading

* add background image and add credit to image creator

* hard reset css styles by making padding and mmargin 0, and updating box-sizing. added background image

however, not yet visible

* make main entire screen. removed flexbox.

my thinking at this stage is make the entire image the 'background', then use z-index to bring .right up so it's on top of the background image in the foreground. not sure if this will work but worth a try. i also used dvh instead of vh because of stackoverflow answer here: https://stackoverflow.com/questions/74144034/why-is-the-css-height100vh-rule-exceeding-the-viewport-height-on-mobile-device so that the height auto-adjusts for mobile state

* show entirety of background image

removed width: 100% as this was unnecessary. used background-size: cover to achieve full visibility

* update background image positioning to show the pinnacle of the mountain on the left side

* make background image responsive by using clamp

fix previous issue where a narrower screen immediately causes white spaces above and below

* drop credit to background photo to the bottom to make it a 'sticky footer'

used flexbox trick from css-tricks.com: https://css-tricks.com/couple-takes-sticky-footer/

* adding padding to text

* add odin logo and font

restructured folder for clarity

* add jost font

this is a free alternative to futura, which is a really nice font!

* modify site name

* center logo and text, set max-width of left side like the design requirements, and set min-height to ensure the background image is still shown fully rather than having white space on the top and bottom

* have pinnacle of mountain visible

* increase width of left side

* add background colour to logo and name (HL), and update colour and positioning (centered) of credit

* add padding to site logo and name container

* reduce logo size and add gap

gap aligns with letter-spacing for aesthetics

* make credit for photo small (semantic and also styling)

* position logo and name of business higher

used position: absolute and top adjuster. had to make .content (the container for the .site-name) position: relative for position: absolute to be relative to it (.content becomes the 'ancestor element' required)